+++ build-config/fragments/plugins.js
@@ -6,18 +6,20 @@
 const path = require('path');
 const moment = require('moment');
 const webpack = require('webpack');
-
 const time = moment().format('YYYY');
-const pckg = require(path.join(__dirname, '..', '..', 'package.json'));
+const rootDir = path.resolve(__dirname, '..', '..');
+const pckg = require(path.join(rootDir, 'package.json'));
 
 // plugin banner with copyright and version info
-var bannerPlugin = new webpack.BannerPlugin(
-`[name] plugin for ${pckg.name}
+var jsBanner = `[name] plugin for ${pckg.name}
 @version ${pckg.version}
 @see ${pckg.homepage}
 @copyright 2014-${time} ${pckg.author}
-@license ${pckg.license}`
-);
+@license ${pckg.license}`;
+var jsBannerPlugin = new webpack.BannerPlugin({
+    banner: jsBanner,
+    test: /\.js$/
+});
 
 /**
  * buildPluginEntry
@@ -31,9 +33,8 @@ function buildPluginEntry(plugins) {
     const result = {};
     plugins.forEach(
         plugin =>
-            (result[plugin.split('-plugin')[0]] = path.resolve(
-                __dirname,
-                '..', '..', 'src', 'js', 'plugins', plugin
+            (result[plugin.split('-plugin')[0]] = path.join(
+                rootDir, 'src', 'js', 'plugins', plugin
             ))
     );
     return result;
@@ -47,11 +48,11 @@ module.exports = {
         'recorderjs-plugin'
     ]),
     output: {
-        path: path.resolve(__dirname, '..', '..', 'dist', 'plugins'),
+        path: path.join(rootDir, 'dist', 'plugins'),
         filename: 'videojs.record.[name].js',
         library: ['VideojsRecord', '[name]']
     },
     plugins: [
-        bannerPlugin
+        jsBannerPlugin
     ]
 };