+++ test/controls/camera-button.spec.js
@@ -26,52 +26,53 @@ describe('controls.CameraButton', function() {
         jasmine.DEFAULT_TIMEOUT_INTERVAL = originalTimeout;
     });
 
-    it('should create the correct DOM element', function() {
+    it('creates the correct DOM element', function() {
         let button = new CameraButton(player);
 
         expect(button.el().nodeName).toEqual('BUTTON');
         expect(button.on).toBeFunction();
-        expect(button.enabled_).toEqual(true);
+        expect(button.enabled_).toBeTrue();
         expect(button.controlText_).toEqual('Image');
 
-        ['vjs-camera-button', 'vjs-control', 'vjs-button', 'vjs-icon-photo-camera'].forEach(
-        (e) => {
-            expect(button.hasClass(e)).toEqual(true);
+        let styleClasses = ['vjs-camera-button', 'vjs-control', 'vjs-button',
+            'vjs-icon-photo-camera'];
+        styleClasses.forEach((e) => {
+            expect(button.hasClass(e)).toBeTrue();
         });
     });
 
-    it('should disable', function() {
+    it('can be disabled', function() {
         let button = new CameraButton(player);
         button.disable();
         expect(button.enabled_).toBeFalse();
     });
 
-    it('should change appearance when startRecord or stopRecord is triggered', function(done) {
+    it('changes appearance when startRecord or stopRecord is triggered', function(done) {
         let button = new CameraButton(player);
 
-        expect(button.hasClass('vjs-icon-photo-camera')).toEqual(true);
+        expect(button.hasClass('vjs-icon-photo-camera')).toBeTrue();
 
         player.one('startRecord', function() {
-            expect(button.hasClass('vjs-icon-photo-camera')).toEqual(false);
-            expect(button.hasClass('vjs-icon-replay')).toEqual(true);
+            expect(button.hasClass('vjs-icon-photo-camera')).toBeFalse();
+            expect(button.hasClass('vjs-icon-replay')).toBeTrue();
             expect(button.controlText_).toEqual('Retry');
 
             setTimeout(function() {
                 // stop recording
-                player.record().stop();
+                button.trigger('click');
 
                 done();
             }, 2000);
         });
         player.one('stopRecord', function() {
-            expect(button.hasClass('vjs-icon-replay')).toEqual(false);
-            expect(button.hasClass('vjs-icon-photo-camera')).toEqual(true);
+            expect(button.hasClass('vjs-icon-replay')).toBeFalse();
+            expect(button.hasClass('vjs-icon-photo-camera')).toBeTrue();
             expect(button.controlText_).toEqual('Image');
 
             done();
         });
         player.one('deviceReady', function() {
-            player.record().start();
+            button.trigger('click');
         });
 
         player.one('ready', function() {