+++ test/controls/record-canvas.spec.js
@@ -8,12 +8,12 @@ import RecordCanvas from '../../src/js/controls/record-canvas.js';
 /** @test {record-canvas} */
 describe('controls.RecordCanvas', function() {
 
-    it('should create the correct DOM element', function() {
+    it('creates the correct DOM element', function() {
         let canvas = new RecordCanvas();
 
         expect(canvas.el().nodeName).toEqual('DIV');
         expect(canvas.on).toBeFunction();
         expect(canvas.hasClass('vjs-record-canvas')).toBeTrue();
-        expect(canvas.el().innerHTML).toEqual('<canvas></canvas>')
+        expect(canvas.el().innerHTML).toEqual('<canvas></canvas>');
     });
 });