+++ test/controls/record-toggle.spec.js
@@ -20,60 +20,61 @@ describe('controls.RecordToggle', function() {
         player.dispose();
     });
 
-    it('should create the correct DOM element', function() {
+    it('creates the correct DOM element', function() {
         let toggle = new RecordToggle(player);
 
         expect(toggle.el().nodeName).toEqual('BUTTON');
         expect(toggle.on).toBeFunction();
-        expect(toggle.enabled_).toEqual(true);
+        expect(toggle.enabled_).toBeTrue();
         expect(toggle.controlText_).toEqual('Record');
 
-        ['vjs-record-button', 'vjs-control', 'vjs-button', 'vjs-icon-record-start'].forEach(
-        (e) => {
-            expect(toggle.hasClass(e)).toEqual(true);
+        let styleClasses = ['vjs-record-button', 'vjs-control', 'vjs-button',
+            'vjs-icon-record-start'];
+        styleClasses.forEach((e) => {
+            expect(toggle.hasClass(e)).toBeTrue();
         });
     });
 
-    it('should disable', function(done) {
+    it('can be disabled', function(done) {
         let toggle = new RecordToggle(player);
 
         player.one('ready', function() {
             toggle.disable();
-            expect(toggle.enabled_).toEqual(false);
+            expect(toggle.enabled_).toBeFalse();
 
             done();
-        })
+        });
     });
 
-    it('should change appearance when startRecord or stopRecord is triggered', function(done) {
+    it('change appearance when startRecord or stopRecord is triggered', function(done) {
         let toggle = new RecordToggle(player);
 
-        expect(toggle.hasClass('vjs-icon-record-start')).toEqual(true);
+        expect(toggle.hasClass('vjs-icon-record-start')).toBeTrue();
 
         player.one('ready', function() {
             player.trigger('startRecord');
 
-            expect(toggle.hasClass('vjs-icon-record-start')).toEqual(false);
-            expect(toggle.hasClass('vjs-icon-record-stop')).toEqual(true);
+            expect(toggle.hasClass('vjs-icon-record-start')).toBeFalse();
+            expect(toggle.hasClass('vjs-icon-record-stop')).toBeTrue();
             expect(toggle.controlText_).toEqual('Stop');
 
             player.trigger('stopRecord');
 
-            expect(toggle.hasClass('vjs-icon-record-stop')).toEqual(false);
-            expect(toggle.hasClass('vjs-icon-record-start')).toEqual(true);
+            expect(toggle.hasClass('vjs-icon-record-stop')).toBeFalse();
+            expect(toggle.hasClass('vjs-icon-record-start')).toBeTrue();
             expect(toggle.controlText_).toEqual('Record');
 
             done();
         });
     });
 
-    it('should accept interaction', function(done) {
+    it('accept interaction', function(done) {
         let toggle = new RecordToggle(player);
 
         player.one('deviceReady', function() {
             // start
             toggle.trigger('click');
-            expect(player.record().isRecording()).toEqual(true);
+            expect(player.record().isRecording()).toBeTrue();
 
             done();
         });