+++ test/engine/record-engine.spec.js
@@ -20,16 +20,16 @@ describe('engine.record-engine', function() {
         player.dispose();
     });
 
-    it('should create the correct component', function() {
+    it('create the correct component', function() {
         let engine = new RecordEngine(player, {});
 
         expect(engine.on).toBeFunction();
-        
+
         // should auto mixin the evented mixin (required since video.js v6.6.0)
         expect(engine.options_.evented).toBeTrue();
     });
 
-    it('should contain supported recorder plugin engines', function() {
+    it('contain supported recorder plugin engines', function() {
         expect(RECORDRTC).toEqual('recordrtc');
         expect(LIBVORBISJS).toEqual('libvorbis.js');
         expect(RECORDERJS).toEqual('recorder.js');
@@ -37,23 +37,24 @@ describe('engine.record-engine', function() {
         expect(OPUSRECORDER).toEqual('opus-recorder');
     });
 
-    it('should trigger recordComplete event', function(done) {
+    it('trigger recordComplete event', function(done) {
         let engine = new RecordEngine(player, {});
         engine.on('recordComplete', function() {
             done();
-        })
+        });
 
         let data = {};
         engine.onStopRecording(data);
     });
 
-    it('should add file info', function(done) {
+    it('add file info', function(done) {
         let engine = new RecordEngine(player, {});
         engine.on('recordComplete', function() {
-            expect(engine.recordedData.name).toEqual(
-                   engine.recordedData.lastModified + '.ogg');
+            let fileName = engine.recordedData.lastModified + '.ogg';
+            expect(engine.recordedData.name).toEqual(fileName);
+
             done();
-        })
+        });
 
         let req = new Request(TestHelpers.TEST_OGG);
         fetch(req).then(function(response) {
@@ -63,7 +64,7 @@ describe('engine.record-engine', function() {
         });
     });
 
-    it('should save as', function(done) {
+    it('save as', function(done) {
         let engine = new RecordEngine(player, {});
         engine.on('recordComplete', function() {
             let fileName = 'foo';
@@ -72,7 +73,7 @@ describe('engine.record-engine', function() {
             let element = document.getElementsByTagName('a')[0];
             expect(element.download).toEqual(fileName);
             done();
-        })
+        });
 
         let req = new Request(TestHelpers.TEST_OGG);
         fetch(req).then(function(response) {